Moving Document Builder files
