+++ src/components/documentBuilder/documentTree.tsx
@@ -16,7 +16,7 @@
  */
 
 import React from "react";
-import BlockPipeline from "@/components/documentBuilder/DocumentBlock";
+import BlockPipeline from "@/components/documentBuilder/render/DocumentBlock";
 import { isExpression, isPipelineExpression } from "@/runtime/mapArgs";
 import { UnknownObject } from "@/types";
 import { get } from "lodash";
@@ -27,11 +27,11 @@ import {
   DocumentElement,
   PipelineDocumentConfig,
 } from "./documentBuilderTypes";
-import DocumentButton from "@/components/documentBuilder/DocumentButton";
+import DocumentButton from "@/components/documentBuilder/render/DocumentButton";
 import useNotifications from "@/hooks/useNotifications";
 import documentTreeStyles from "./documentTree.module.scss";
 import cx from "classnames";
-import DocumentList from "@/components/documentBuilder/DocumentList";
+import DocumentList from "@/components/documentBuilder/render/DocumentList";
 
 const headerComponents = {
   header_1: "h1",
