+++ src/components/documentBuilder/edit/MoveElementAction.tsx
@@ -19,7 +19,7 @@ import { useField } from "formik";
 import React from "react";
 import FieldTemplate from "@/components/form/FieldTemplate";
 import getElementCollectionName from "./getElementCollectionName";
-import { DocumentElement } from "./documentBuilderTypes";
+import { DocumentElement } from "@/components/documentBuilder/documentBuilderTypes";
 import LayoutWidget from "@/components/LayoutWidget";
 
 type MoveElementActionProps = {
