+++ src/components/documentBuilder/preview/AddElementAction.tsx
@@ -22,9 +22,9 @@ import {
   DocumentElement,
   DocumentElementType,
   DOCUMENT_ELEMENT_TYPES,
-} from "./documentBuilderTypes";
+} from "@/components/documentBuilder/documentBuilderTypes";
 import styles from "./AddElementAction.module.scss";
-import { createNewElement } from "./createNewElement";
+import { createNewElement } from "@/components/documentBuilder/createNewElement";
 import elementTypeLabels from "@/components/documentBuilder/elementTypeLabels";
 
 type AddElementActionProps = {
