+++ src/components/documentBuilder/preview/ElementPreview.tsx
@@ -18,11 +18,14 @@
 import React, { MouseEventHandler } from "react";
 import styles from "./ElementPreview.module.scss";
 import cx from "classnames";
-import { DocumentElement, isListDocument } from "./documentBuilderTypes";
-import { getPreviewComponentDefinition } from "./documentTree";
+import {
+  DocumentElement,
+  isListDocument,
+} from "@/components/documentBuilder/documentBuilderTypes";
+import { getPreviewComponentDefinition } from "@/components/documentBuilder/documentTree";
 import AddElementAction from "./AddElementAction";
 import { useField } from "formik";
-import { getAllowedChildTypes } from "./allowedElementTypes";
+import { getAllowedChildTypes } from "@/components/documentBuilder/allowedElementTypes";
 
 interface ElementPreviewTemplateProps {
   elementName: string;
