+++ src/components/documentBuilder/render/DocumentButton.tsx
@@ -1,10 +1,27 @@
+/*
+ * Copyright (C) 2021 PixieBrix, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 import React, { useCallback, useContext } from "react";
 import { BlockPipeline } from "@/blocks/types";
 import AsyncButton, { AsyncButtonProps } from "@/components/AsyncButton";
 import { whoAmI } from "@/background/messenger/api";
 import { runEffectPipeline } from "@/contentScript/messenger/api";
 import { uuidv4 } from "@/types/helpers";
-import DocumentContext from "@/components/documentBuilder/DocumentContext";
+import DocumentContext from "@/components/documentBuilder/render/DocumentContext";
 import { Except } from "type-fest";
 
 type DocumentButtonProps = Except<AsyncButtonProps, "onClick"> & {
